Avoid NPE when cleaning product status.

+++ src/main/java/ecmwf/ecpds/master/plugin/http/dao/monitoring/ProductStepStatusDAOHandler.java
@@ -179,7 +179,7 @@ public class ProductStepStatusDAOHandler extends ReadOnlyDAOHandler implements D
             while (i.hasNext()) {
                 final var key = i.next().toString();
                 final var pss = (ProductStepStatus) CacheService.get(klass, key);
-                if (product.equals(pss.getProduct()) && time.equals(pss.getTime())) {
+                if (pss != null && product.equals(pss.getProduct()) && time.equals(pss.getTime())) {
                     // Completely DELETE this entry as the same step might not be available for next iteration
                     CacheService.remove(klass, key);
                 }
